+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.0] - 2022-03-03
 ### Added
 - Tests for "Blowing fuse".
 - Tests for "Container terminal".
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Rectangle partition".
 - Tests for "Robot show".
 - Tests for "Rock paper scissors lizard Spock".
+- Tests for "Rooks movements".
 
 ## [1.4.0] - 2022-03-03
 ### Added